Refactor compiler package
The new Compiler trait now has DOT types that declare how a lang is to be implemented: a monadic context for use in a dsl to build up a correct-by-construction ast, a definition level for introducing defs or bare expressions into the context, an entity level to represent the sorts of identifiers that will be defined and/or referred to, and an error type. Then it is easy for implementers to define compile, toString, toStringOrError, and toVirtualFile.